API now exists in its own module file. Stubs for other endpoints, added versioning to url.
